apps/api/pyproject.toml
[project]
name = "autointel-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.21",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "opensearch-py>=2.5.0",
  "boto3>=1.34.0",
  "redis>=5.0.0",
  "rq>=1.16.0",
  "numpy>=2.0.0",
  "pandas>=2.2.0",
  "joblib>=1.4.0",
  "scikit-learn>=1.8.0",
  "nbclient>=0.10.0",
  "nbformat>=5.10.0",
  "ipykernel>=6.29.0",
  "pillow>=10.4.0",
  "pytesseract>=0.3.13",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*"]
namespaces = true
